+++ geopackage-ios/geom/map/GPKGMapShapePoints.h
@@ -162,6 +162,21 @@
 +(void) addPointAsPolyline: (GPKGMapPoint *) point toPoints: (NSMutableArray *) points;
 
 /**
+ *  Updates hidden state of all objects
+ *
+ *  @param hidden hidden flag
+ *  @param mapView map view
+ */
+-(void) hidden: (BOOL) hidden fromMapView: (MKMapView *) mapView;
+
+/**
+ *  Updates hidden state of the shape representing points
+ *
+ *  @param hidden hidden flag
+ */
+-(void) hiddenPoints: (BOOL) hidden;
+
+/**
  * Get the shape markers size
  *
  * @return size